fix user selector component
